Ai update grok

See merge request dn-quest/quiz-engine!16

+++ quest-management-service/src/main/java/dn/quest/questmanagement/event/QuestDeletedEvent.java
@@ -17,7 +17,7 @@ public class QuestDeletedEvent extends QuestEvent {
     public QuestDeletedEvent(Long questId, Long questNumber, String title, String description,
                             String difficulty, String questType, String category,
                             java.util.Set<String> tags, java.util.Set<Long> authorIds,
-                            String status, Integer version, Long userId) {
+                            String status, Long version, Long userId) {
         super("QUEST_DELETED", questId, userId);
         this.setQuestNumber(questNumber);
         this.setTitle(title);
