+++ quest-management-service/src/main/java/dn/quest/questmanagement/event/QuestUpdatedEvent.java
@@ -17,7 +17,7 @@ public class QuestUpdatedEvent extends QuestEvent {
     public QuestUpdatedEvent(Long questId, Long questNumber, String title, String description,
                            String difficulty, String questType, String category,
                            java.util.Set<String> tags, java.util.Set<Long> authorIds,
-                           String status, Integer version, Long userId) {
+                           String status, Long version, Long userId) {
         super("QUEST_UPDATED", questId, userId);
         this.setQuestNumber(questNumber);
         this.setTitle(title);
